corechecks/network: split networkv2 into common + per-OS files
Move the pieces that are byte-identical between linux and windows
(CheckName, the NetworkCheck/networkConfig/networkInitConfig structs,
Factory) into network.go, leaving network_linux.go / network_windows.go
with everything else, unchanged, just relocated. Everything that isn't
identical between the two platforms (Configure's tail, device-exclusion
checks, submitInterfaceMetrics, submitConnectionsMetrics) is left exactly
as it was, duplicated, to keep this diff a pure move rather than a rewrite.

Meant to make it easier to add AIX support (as its own network_aix.go) in a
follow-up PR without repeating this scaffolding a third time.

Validated: full Linux test suite (55/55) and linter unmodified; GOOS=windows
cross-lint (0 issues); confirmed AIX still builds (falls through to the
existing no-op stub, unchanged from before this refactor).
